extract point into a new file
